requirements.txt
numpy~=1.25.1
pandas~=2.0.3
rdkit~=2023.3.2
dask~=2023.7.1
torch~=2.0.1
biopython~=1.81